requirements.txt
# Core dependencies
ollama>=0.1.7
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
opencv-python>=4.8.0
spacy>=3.7.0
nltk>=3.8.1
numpy>=1.21.6,<1.28.0

# Windows-specific dependencies
pywin32>=306
pywin32-ctypes>=0.2.2
pyinstaller>=6.0.0

# UI and notifications
plyer>=2.1.0
tkinter-tooltip>=2.1.0

# Utilities
requests>=2.31.0
pathlib2>=2.3.7
python-dotenv>=1.0.0
psutil>=5.9.0
watchdog>=3.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Logging and monitoring
colorlog>=6.7.0
python-json-logger>=2.0.7

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Document processing
python-docx>=0.8.11
striprtf>=0.0.26

# Configuration management
pyyaml>=6.0.1
configparser>=6.0.0

# Security
cryptography>=41.0.0